requirements.txt
pillow-heif==0.21.0
Pillow==10.2.0
python-webp
